[Ruby] Use nil as default value for all types (#14)
The constructors were setting an empty string or the number zero
depending on the type of the attribute. This made it impossible to tell
if the value was actively set to zero or if it was not set at all.

Now it's possible to omit the stockLimitation attribute for unlimited
stock.

This change also reduces the JSON file size by omitting these values.