Upgrade jsonld-streaming-parser to v4.
The primary reason for upgrading jsonld-streaming-parser to v4 is to
upgrade jsonld-context-parser to v3, because in 2.4.0 the `load` method
on `FetchDocumentLoader` failed to redirect correctly to the alternate
URL for DFC's context JSON-LD, which does seem to be correctly included
as a `"link"`` header for www.datafoodconsortium.org. Specifically, it
caused several tests to fail whenever the `Connector.import()` method
was called. It may be that this could have been occurring earlier but
only surfaced after a recent change to the DFC site or the context URL.
The current URL for DFC's context JSON-LD document is:

https://www.datafoodconsortium.org/wp-content/plugins/wordpress-context-jsonld/context.jsonld

Looking at the diffs between 2.4.0 and 3.0.0, the culprit may have been
the removal of the cross-fetch polyfill, but that's just a guess. See
relevant code, commits, and diffs from the `jsonld-context-parser` repo:

https://github.com/rubensworks/jsonld-context-parser.js/blob/v3.0.0/lib/FetchDocumentLoader.ts
https://github.com/rubensworks/jsonld-context-parser.js/compare/v2.3.0..v3.0.0
https://github.com/rubensworks/jsonld-context-parser.js/commit/8d26fba6cb2e8c54f998a2eafeee49a4688abe36